Remove configuration object.
After sleeping on it, makes more sense to leverage a method in
ApplicationMarkdown because an application can set up multiple renderers
and it would be good to have the option to configure those on a
per-renderer basis.

This updates the main module to remove the configuration, and adds
callouts in the ApplicationMarkdown template to use this feature.